scripts/acoustic_distance/requirements.txt
dtw-python
editdistance
matplotlib
numpy
onnxruntime
pandas
scipy
seaborn
SoundFile
torch
transformers
tqdm